pyproject.toml
[project]
name = "nigerian-tax-tracker"
version = "0.1.0"
description = "Tax tracking application for Nigerian LLCs"
authors = [
    {name = "Olalekan", email = "olalekan@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "reportlab>=4.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "aiofiles>=23.0.0",
    "pdfplumber>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
